mcp/qdrant_rag_server/requirements.txt
qdrant-client>=1.9.0
python-dotenv>=1.0.0